almost match RFL_NANDAccess.c

+++ include/context_sdk/revolution/mem/heapHandle.h
@@ -9,7 +9,7 @@ extern "C" {
 #endif
 
 typedef struct MEMiHeapHead {
-    unsigned long signature; // 0x0
+    u32 signature; // 0x0
     struct {
         void * prevObject; // 0x0
         void * nextObject; // 0x4
@@ -17,17 +17,17 @@ typedef struct MEMiHeapHead {
     struct {
         void * headObject; // 0x0
         void * tailObject; // 0x4
-        unsigned short numObjects; // 0x8
-        unsigned short offset; // 0xA
+        u16 numObjects; // 0x8
+        u16 offset; // 0xA
     } childList; // 0xC
     void * heapStart; // 0x18
     void * heapEnd; // 0x1C
     struct OSMutex mutex; // 0x20
     union {
-        unsigned long val; // 0x0
+        u32 val; // 0x0
         struct {
-            unsigned long _reserved : 24; // 0x0
-            unsigned long optFlag : 8; // 0x0
+            u32 _reserved : 24; // 0x0
+            u32 optFlag : 8; // 0x0
         } fields; // 0x0
     } attribute; // 0x38
 } MEMiHeapHead;
